show every card the last player played instead of just the last
Before this change, this caused a weird bug when the AI played two cards
in a row, it could sometimes look like they broke the rules. For
example, if the last card played was a red skip, and the ai played first
a blue skip then a blue 3, to the next player it would look like the ai
played a blue 3 on top of a red skip, which is breaking the rules.